requirements.txt
discord.py
gspread
oauth2client
discord.py
gspread
oauth2client
flask
